// lib/types.ts
import { z } from "zod"
import {ConnectionProviderProps} from "@/providers/connections-provider";

export const EditUserProfileSchema = z.object({
    email : z.string().email("Required"),
    name : z.string().min(1,"Name is required"),
})

export type ConnectionTypes = 'Google Drive' | 'Notion' | 'Slack' | 'Discord'

export type Connection = {
    title: ConnectionTypes
    description: string
    image: string
    connectionKey: keyof ConnectionProviderProps
    accessTokenKey?: string
    alwaysTrue?: boolean
    slackSpecial?: boolean
}